refactor(schema): extract some code to use it in full hosted context
